day03/requirements.txt
streamlit>=1.35.0
google-generativeai>=0.7.0
pydantic>=2.0.0
python-dotenv>=1.0.0
